Move test and WildcardEnumPojo to more appropriate packages
